+++ src/GammonX/GammonX.Server/Models/gameSession/BackgammonGameSession.cs
@@ -4,25 +4,31 @@ using GammonX.Models.Enums;
 
 namespace GammonX.Server.Models.gameSession
 {
-	public interface IDoublingCubeGameSession
+    /// <summary>
+    /// Provides the capabilities for a backgammon game session, including handling of the doubling cube actions.
+    /// </summary>
+    public interface IDoublingCubeGameSession
 	{
 		/// <summary>
 		/// Changes the active player without any other effect (e.g. turnnumber)
 		/// </summary>
 		/// <param name="callingPlayerId">Player who offered the double.</param>
-		void DoubleOffered(Guid callingPlayerId);
+		/// <param name="isWhite">Indicates if the player is white.</param>
+		void DoubleOffered(Guid callingPlayerId, bool isWhite);
 
-		/// <summary>
-		/// Changes the active player without any other effect (e.g. turnnumber)
-		/// </summary>
-		/// <param name="callingPlayerId">Player who acceppted the double.</param>
-		void DoubleAccepted(Guid callingPlayerId);
+        /// <summary>
+        /// Changes the active player without any other effect (e.g. turnnumber)
+        /// </summary>
+        /// <param name="callingPlayerId">Player who acceppted the double.</param>
+        /// <param name="isWhite">Indicates if the player is white.</param>
+        void DoubleAccepted(Guid callingPlayerId, bool isWhite);
 
-		/// <summary>
-		/// Changes the active player without any other effect (e.g. turnnumber)
-		/// </summary>
-		/// <param name="callingPlayerId">Player who declined the double.</param>
-		void DoubleDeclined(Guid callingPlayerId);
+        /// <summary>
+        /// Changes the active player without any other effect (e.g. turnnumber)
+        /// </summary>
+        /// <param name="callingPlayerId">Player who declined the double.</param>
+        /// <param name="isWhite">Indicates if the player is white.</param>
+        void DoubleDeclined(Guid callingPlayerId, bool isWhite);
 	}
 
 	// <inheritdoc />
@@ -35,22 +41,25 @@ namespace GammonX.Server.Models.gameSession
 		}
 
 		// <inheritdoc />
-		public void DoubleAccepted(Guid callingPlayerId)
+		public void DoubleAccepted(Guid callingPlayerId, bool isWhite)
 		{
-			ActivePlayer = OtherPlayer;
+            BoardService.AddCubeEventToHistory(BoardModel, isWhite, CubeAction.Take);
+            ActivePlayer = OtherPlayer;
 			OtherPlayer = callingPlayerId;
 		}
 
 		// <inheritdoc />
-		public void DoubleDeclined(Guid callingPlayerId)
+		public void DoubleDeclined(Guid callingPlayerId, bool isWhite)
 		{
-			ActivePlayer = OtherPlayer;
+            BoardService.AddCubeEventToHistory(BoardModel, isWhite, CubeAction.Pass);
+            ActivePlayer = OtherPlayer;
 			OtherPlayer = callingPlayerId;
 		}
 
 		// <inheritdoc />
-		public void DoubleOffered(Guid callingPlayerId)
+		public void DoubleOffered(Guid callingPlayerId, bool isWhite)
 		{
+			BoardService.AddCubeEventToHistory(BoardModel, isWhite, CubeAction.Offer);
 			ActivePlayer = OtherPlayer;
 			OtherPlayer = callingPlayerId;
 		}